# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/log/food": {
      "post": {
        "operationId": "LogController_logFood",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogFoodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
    "/log/daily": {
      "get": {
        "operationId": "LogController_getDaily",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
    "/log/food/{id}": {
      "put": {
        "operationId": "LogController_updateEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLogEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      },
      "delete": {
        "operationId": "LogController_deleteEntry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
    "/log/weekly": {
      "get": {
        "operationId": "LogController_getWeekly",
        "parameters": [
          {
            "name": "week_start",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "log"
        ]
      }
    },
...
    "/coach/clients/{client_id}/nudges": {
      "post": {
        "operationId": "CoachNudgesController_create",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNudgeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "nudges"
        ]
      }
    },
...
    "/coach/clients/{client_id}/meal-plans": {
      "get": {
        "operationId": "CoachMealPlansController_list",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "meal-plans"
        ]
      },
      "post": {
        "operationId": "CoachMealPlansController_create",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "meal-plans"
        ]
      }
    },
    "/coach/meal-plans/{id}": {
      "patch": {
        "operationId": "CoachMealPlansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMealPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "meal-plans"
        ]
      },
      "delete": {
        "operationId": "CoachMealPlansController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "meal-plans"
        ]
      }
    },
...
    "/meal-plans/{id}": {
      "get": {
        "operationId": "ClientMealPlansController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "meal-plans"
        ]
      }
    },
...
    "/recipes": {
      "get": {
        "operationId": "RecipesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "recipes"
        ]
      },
      "post": {
        "operationId": "RecipesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "recipes"
        ]
      }
    },
...
    "/recipes/{id}": {
      "get": {
        "operationId": "RecipesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "recipes"
        ]
      }
    },
    "/recipes/{id}/save": {
      "post": {
        "operationId": "RecipesController_save",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "recipes"
        ]
      },
      "delete": {
        "operationId": "RecipesController_unsave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "recipes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth Project API",
    "description": "HTTP API for The Growth Project — coach/client SaaS platform covering auth, billing, messaging, check-ins, meal plans, workouts, habits, and admin/federation surfaces. Consumed by the React Native mobile app, the coach console BFF, and partner integrations.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.trygrowthproject.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase-issued JWT. Send as `Authorization: Bearer <token>`. Tokens are verified against the Supabase JWKS in JwtAuthGuard."
      }
    },
    "schemas": {
...
      "LogFoodDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLogEntryDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateNudgeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateMealPlanDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMealPlanDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRecipeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}